Split ApplicationList into ApplicationList and ApplicationListElement
